Improved structure for the SGP4 Hoots implementation.
The pattern used in the SGP4 Hoots implementation is to perform a
calculation and return some State that is composed of results + some
calculation terms + previous terms.

This would pave the way to introduce monadic composition.

This structure probably contains errors in the formulas. No tests are
done at the moment.