Vallado's SGP4 implemtation shows agreement to 1E-9.
The 3 tests for near Earth TLEs show that the SGP4Vallado
implementation (based on Vallado's code) matches the results given by
Vallado's CPP implementation up to 1E-9 for the position and the
velocity (given in km and m/s).

Note as well that the implicit value for Equality has to be given
explicitly to the function objects for the checks to get the call done.
Otherwise, it remains a function definition Eq[Double] => Unit.